Freeze requirements on deployment
prepare.sh now saves the package versions at the state of deployment of a new commit to alpha/beta/production
into a .frozen-requirements.txt file. As long as the HEAD commit doesn't change, every restart
will install the exact same package versions, to prevent silent, unwanted, breaking upgrades to some degree.

Also pin Flask to <2 and Jinja2 <3 until we tested it.
Prevent SqlAlchemy from upgrading to 2.0 when it is released,
since we use several deprecated functions that will be removed.
Remove some older pins that got added as a result of upstream bugs that should be resolved now.

Change Werkzeug profiler import from `werkzeug.contrib.profiler` to `werkzeug.middleware.profiler`,
it has been moved in 1.0.

+++ prepare.sh
@@ -1,7 +1,31 @@
 #!/bin/bash
+# Compares the current git HEAD with the one of the previous deployment.
+# If it is the same, install the same requirements version as last time.
+# If it differs, upgrade all packages to the latest version.
+
 set -e
 
-# works inside SpaceDock folder on alpha/beta, which is a venv
-bin/pip3 install --no-cache-dir -U -r requirements.txt
+if [[ -f .last_head ]]; then
+    LAST_HEAD=$(cat .last_head)
+else
+    LAST_HEAD=""
+fi
 
-./build-frontend.sh
+HEAD=$(git rev-parse HEAD)
+
+# Make sure we've got the latest pip, to benefit of security and relationship resolver fixes.
+bin/pip3 install -U pip
+
+if [[ "$LAST_HEAD" != "$HEAD" || ! -f .frozen-requirements.txt ]]; then
+  echo "HEAD at last deployment: $LAST_HEAD - Current HEAD: $HEAD"
+  # works inside SpaceDock folder on alpha/beta, which is a venv
+  bin/pip3 install --no-cache-dir -U -r requirements.txt
+  bin/pip3 freeze > .frozen-requirements.txt
+
+  # Frontend only needs to be rebuilt if the commit changed
+  ./build-frontend.sh
+else
+  bin/pip3 install --no-cache-dir -r .frozen-requirements.txt
+fi
+
+echo "$HEAD" > .last_head
